+++ inc/ti/type.h
@@ -10,6 +10,7 @@
 #include <ti/closure.t.h>
 #include <ti/method.t.h>
 #include <ti/name.t.h>
+#include <ti/map.h>
 #include <ti/thing.t.h>
 #include <ti/type.t.h>
 #include <ti/types.t.h>
@@ -40,7 +41,7 @@ int ti_type_init_from_unp(
         _Bool with_hide_id);
 int ti_type_fields_to_pk(ti_type_t * type, msgpack_packer * pk);
 ti_val_t * ti_type_as_mpval(ti_type_t * type, _Bool with_definition);
-vec_t * ti_type_map(ti_type_t * to_type, ti_type_t * from_type);
+ti_map_t * ti_type_map(ti_type_t * to_type, ti_type_t * from_type);
 ti_val_t * ti_type_dval(ti_type_t * type);
 int ti_type_convert(
         ti_type_t * type,