+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.7.5
 
 * Added `match()` function and global (`g`) flag for regular expressions, pr #409.
+* Added `?` prefix flag to exclude `nil` values, pr #411.
 
 # v1.7.4